+++ html-link-parser/parser.go
@@ -1,7 +1,6 @@
 package link
 
 import (
-	"fmt"
 	"golang.org/x/net/html"
 	"io"
 	"strings"
@@ -29,7 +28,7 @@ func Parse(r io.Reader) ([]Link, error) {
 
 	for _, node := range nodes {
 		links = append(links, buildLink(node))
-		fmt.Println(node)
+		//fmt.Println(node)
 	}
 	//dfs(doc, "")
 	return links, nil