+++ main.go
@@ -1,8 +1,16 @@
 package main
 
+import sitemap_builder "goPhercise/sitemap-builder"
+
 func main() {
 	//quiz_game.RollQuizGame()
 	//url_shortener.RollURLShortener()
 	//cya_game.RollCYAGame()
 	//HTML Link parser is run in its own directories
+	//aecho.RollEcho()
+	//bduplines.RollDupLines()
+	//clissajous.RollLissajous()
+	//dfetchurl.RollFetchURLConcurrent()
+	//ewebserver.RollWebServer()
+	sitemap_builder.RollSitemapBuilder()
 }